requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.6

docker-py>=1.6.0,<1.8.0  # Apache-2.0
GitPython>=1.0.1 # BSD License (3 clause)
oslo.config>=3.9.0  # Apache-2.0
oslo.log>=1.14.0  # Apache-2.0
